Module Sihl_web.Form_parser

val log_src : Logs.src

module Logs : Logs.LOG

type body = (string * string list) list

val body_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> body

val sexp_of_body : body -> Ppx_sexp_conv_lib.Sexp.t

exception Parsed_body_not_found

val key : body Opium.Context.key

val find_all : Opium.Request.t -> body

val find : Stdlib.String.t -> Opium.Request.t -> string option

val consume : Opium.Request.t -> Stdlib.String.t -> Opium.Request.t * string option

consume req key returns the value of the parsed body for the key key and a request with an updated context where the parsed value is missing the key key. The value is returned and removed from the context, it is consumed. *

val m : unit -> Rock.Middleware.t