Module Sihl_web__Middleware_urlencoded

val log_src : Logs.src

module Logs : Logs.LOG

type urlencoded = (string * string list) list

val urlencoded_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> urlencoded

val sexp_of_urlencoded : urlencoded -> Ppx_sexp_conv_lib.Sexp.t

exception Urlencoded_not_found

val key : urlencoded Opium.Context.key

val find_all : Opium.Request.t -> urlencoded

val find : Stdlib.String.t -> Opium.Request.t -> string option

val consume : Opium.Request.t -> Stdlib.String.t -> Opium.Request.t * string option

consume req key returns the value of the parsed urlencoded body for the key key and a request with an update context where the parsed urlencoded is missing the key key. The urlencoded value is returned and removed from the context, it is consumed. *

val m : unit -> Rock.Middleware.t